checkpoint(mfusg): USG-T 2.7 authoring + round-trip work in progress
Snapshot of the in-progress USG-T 2.7 support before adding the Priority-1 packages (SGB, QRT, MfUsgDrt, BCF/LPF TABRICH). All changes below are green under autotest/test_usg_transport.py (35 synthetic tests, Ex* real-model loads excluded).

- Boundary list packages CHD/RIV/GHB/DRN: node-based, 0-based internal / 1-based file, AUX transport concentrations, ITMP/-1 reuse.
- BAS/WEL/CLN/DPF/HFB/TIB/ETS: semantic + from-scratch authoring coverage.
- TVM2 semantic rewrite; density-coupled round-trip fixes (CHD/RIV float64, DDF ithickav, BAS CONVERGE).
- NAM external input-path preservation / output rebase (flopy/modflow/mf.py).
- Docs: USGT_roadmap.md, USGT_improvements.md, USGT_PACKAGE_BACKLOG.md, USGT_AGENT_BRIEF.md.

+++ flopy/mfusg/mfusgbct.py
@@ -812,7 +812,7 @@ class MfUsgBct(Package):
         # item 2 ICBUND
         if kwargs["icbndflg"] == 0:
             kwargs["icbund"] = cls._load_prop_arrays(
-                f_obj, model, nlay, np.float32, "icbund", ext_unit_dict
+                f_obj, model, nlay, np.int32, "icbund", ext_unit_dict
             )
 
         # item 3 PRSITY
